requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.30

# Vector Store & Embeddings
faiss-cpu>=1.7.0
tiktoken>=0.5.0

# Document Processing
pypdf>=4.0.0
markdown>=3.5

# Keyword Search
rank-bm25>=0.2.2

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Chinese Text Processing
jieba>=0.42.1

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
tqdm>=4.65.0

# Web Interface
streamlit>=1.30.0

